requirements.txt
# Core Django (Required for Render)
Django==5.0.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
psycopg2-binary==2.9.9

# PostgreSQL Vector Extension (Required)
pgvector==0.2.4

# Web Server (Required for Render)
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0

# Caching & Async Tasks (Required)
redis==5.0.1
celery==5.3.6
django-redis==5.4.0

# Authentication & Security (Required)
djangorestframework-simplejwt==5.3.1
cryptography==41.0.7

# Basic Image Processing (Required)
Pillow==10.2.0

# Utilities (Required)
python-dotenv==1.0.0
requests==2.31.0

# Optional: AI/ML for Image Embeddings (Install separately if needed)
# torch==2.5.1
# torchvision==0.20.1
# transformers==4.36.2
# sentence-transformers==2.2.2
# git+https://github.com/openai/CLIP.git

# Optional: Advanced Features
# textblob==0.17.1
# vaderSentiment==3.3.2
# numpy==1.26.3
# pandas==2.1.4
# scikit-learn==1.3.2
# boto3==1.34.21
# django-storages==1.14.2
